refactor: created new Legend label component
