+++ src/js/components/legend/LegendInfo.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const LegendInfo = ({ info }: { info: string }) => {
+  return (
+    <div className="legend-info">
+      <span>{info}</span>
+    </div>
+  );
+};
+
+export default LegendInfo;
